@@ -31,12 +31,14 @@
         <dep.docker-java.version>3.2.12</dep.docker-java.version>
         <dep.coral.version>2.0.55</dep.coral.version>
         <dep.confluent.version>5.5.2</dep.confluent.version>
-        <project.build.targetJdk>11</project.build.targetJdk>
+        <project.build.targetJdk>17</project.build.targetJdk>
         <air.java.version>1.8.0-202</air.java.version>
-        <dep.trino.version>375</dep.trino.version>
-        <dep.trino.airlift.version>214</dep.trino.airlift.version>
-        <maven.compiler.source>16</maven.compiler.source>
-        <maven.compiler.target>16</maven.compiler.target>
+        <dep.trino.version>420</dep.trino.version>
+        <dep.trino.airlift.version>233</dep.trino.airlift.version>
+        <dep.packaging.version>${dep.trino.airlift.version}</dep.packaging.version>
+        <dep.arrow.version>11.0.0</dep.arrow.version>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
 
         <dep.docker.images.version>53</dep.docker.images.version>
 
@@ -245,8 +455,8 @@
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <configuration combine.self="override">
-                    <source>11</source>
-                    <target>11</target>
+                    <source>17</source>
+                    <target>17</target>
                     <fork>false</fork>
                 </configuration>
             </plugin>
