@@ -4,7 +4,7 @@
 
 package com.vastdata.client;
 
-import javax.inject.Qualifier;
+import com.google.inject.BindingAnnotation;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
@@ -16,7 +16,7 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 @Retention(RUNTIME)
 @Target({FIELD, PARAMETER, METHOD})
-@Qualifier
+@BindingAnnotation
 public @interface ForVast
 {
 }
