@@ -17,6 +17,8 @@ import com.vastdata.client.schema.ImportDataFile;
 import com.vastdata.client.tx.VastTraceToken;
 import com.vastdata.client.tx.VastTransaction;
 import org.mockito.Mock;
+import org.testng.annotations.AfterTest;
+import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
 import java.net.URI;
@@ -24,9 +26,9 @@ import java.nio.charset.StandardCharsets;
 import java.util.Optional;
 import java.util.function.Supplier;
 
-import static org.mockito.Matchers.any;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
-import static org.mockito.MockitoAnnotations.initMocks;
+import static org.mockito.MockitoAnnotations.openMocks;
 
 public class TestImportDataExecutor
 {
@@ -40,6 +42,21 @@ public class TestImportDataExecutor
     URI uri1 = URI.create("http://localhost:8080");
     URI uri2 = URI.create("http://127.0.0.1:8080");
 
+    private AutoCloseable autoCloseable;
+
+    @BeforeTest
+    public void setup()
+    {
+        autoCloseable = openMocks(this);
+    }
+
+    @AfterTest
+    public void tearDown()
+            throws Exception
+    {
+        autoCloseable.close();
+    }
+
     @Test(expectedExceptions = VastRuntimeException.class, expectedExceptionsMessageRegExp = ".*Number of attempts exceeded configuration: " + (NUMBER_OF_RETRIES + 1))
     public void testExecuteThrowsExceptionOnMaxRetries()
             throws VastException
@@ -47,14 +64,14 @@ public class TestImportDataExecutor
         testExecute(503);
     }
 
-    @Test(expectedExceptions = ImportDataFailure.class, expectedExceptionsMessageRegExp = ".*" + ImportDataFailure.REQUEST_EXECUTION_ERROR_NOT_FOUND)
+    @Test(expectedExceptions = ImportDataFailure.class, expectedExceptionsMessageRegExp = ".*" + ImportDataFailure.REQUEST_EXECUTION_ERROR_NOT_FOUND + ".*")
     public void testExecuteThrowsExceptionOnObjectNotFound()
             throws VastException
     {
         testExecute(404);
     }
 
-    @Test(expectedExceptions = ImportDataFailure.class, expectedExceptionsMessageRegExp = ".*" + ImportDataFailure.REQUEST_EXECUTION_ERROR_PERMISSIONS)
+    @Test(expectedExceptions = ImportDataFailure.class, expectedExceptionsMessageRegExp = ".*" + ImportDataFailure.REQUEST_EXECUTION_ERROR_PERMISSIONS + ".*")
     public void testExecuteThrowsExceptionOnBadPermissions()
             throws VastException
     {
@@ -64,9 +81,10 @@ public class TestImportDataExecutor
     private void testExecute(int returnCode)
             throws VastException
     {
-        initMocks(this);
         when(mockTraceToken.toString()).thenReturn("SomeTraceToken");
         when(mockResponse.getStatus()).thenReturn(returnCode);
+        when(mockResponse.getRequestUri()).thenReturn(uri1);
+        when(mockResponse.getErrorMessage()).thenReturn(Optional.empty());
         when(mockResponse.getBytes()).thenReturn("dummy".getBytes(StandardCharsets.UTF_8));
         when(mockClient.importData(any(VastTransaction.class), any(VastTraceToken.class), any(ImportDataContext.class), any(), any(URI.class)))
                 .thenReturn(mockResponse);
