@@ -6,7 +6,11 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.vastdata.client.VastClient;
 import com.vastdata.client.VastConfig;
+import com.vastdata.client.error.VastRuntimeException;
+import com.vastdata.client.error.VastServerException;
 import com.vastdata.trino.VastColumnHandle;
+import com.vastdata.trino.VastTableHandle;
+import com.vastdata.trino.VastTrinoDependenciesFactory;
 import io.trino.spi.connector.ColumnHandle;
 import io.trino.spi.statistics.ColumnStatistics;
 import io.trino.spi.statistics.Estimate;
@@ -14,6 +18,8 @@ import io.trino.spi.statistics.TableStatistics;
 import org.apache.arrow.vector.types.pojo.ArrowType;
 import org.apache.arrow.vector.types.pojo.Field;
 import org.mockito.Mock;
+import org.testng.annotations.AfterTest;
+import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 
 import java.net.URI;
@@ -22,10 +28,14 @@ import java.util.Optional;
 import static com.vastdata.client.importdata.VastImportDataMetadataUtils.BIG_CATALOG_SCHEMA_PREFIX;
 import static com.vastdata.client.importdata.VastImportDataMetadataUtils.BIG_CATALOG_TABLE_NAME;
 import static java.lang.String.format;
-import static org.mockito.Matchers.anyString;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.spy;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
-import static org.mockito.MockitoAnnotations.initMocks;
+import static org.mockito.MockitoAnnotations.openMocks;
 import static org.testng.Assert.assertEquals;
+import static org.testng.Assert.assertTrue;
 
 public class TestTrinoPersistentStatistics
 {
@@ -35,8 +45,24 @@ public class TestTrinoPersistentStatistics
             .setNullsFraction(Estimate.unknown())
             .setRange(new io.trino.spi.statistics.DoubleRange(1.00, 100000.00))
             .build();
+    public static final int RETRY_MAX_COUNT = 1;
     @Mock
     private VastClient mockClient;
+
+    private AutoCloseable autoCloseable;
+
+    @BeforeTest
+    public void setup()
+    {
+        autoCloseable = openMocks(this);
+    }
+
+    @AfterTest
+    public void tearDown()
+            throws Exception
+    {
+        autoCloseable.close();
+    }
 
     private VastConfig getMockServerReadyVastConfig()
     {
@@ -45,21 +71,21 @@ public class TestTrinoPersistentStatistics
                 .setRegion("us-east-1")
                 .setAccessKeyId("pIX3SzyuQVmdrIVZnyy0")
                 .setSecretAccessKey("5c5HqW3cDQsUNg68OlhJmq72TM2nZxcP5lR6D1ps")
-                .setRetryMaxCount(1)
+                .setRetryMaxCount(RETRY_MAX_COUNT)
                 .setRetrySleepDuration(1);
     }
 
     @Test
     public void testSetTableStatistics()
     {
-        initMocks(this);
         VastClient client = mockClient;
         VastConfig config = getMockServerReadyVastConfig();
         TrinoPersistentStatistics persistentStatistics = new TrinoPersistentStatistics(client, config);
-        String url = "/" + BIG_CATALOG_SCHEMA_PREFIX + BIG_CATALOG_TABLE_NAME;
+        VastTableHandle table = new VastTableHandle(BIG_CATALOG_SCHEMA_PREFIX + BIG_CATALOG_SCHEMA_PREFIX,
+                BIG_CATALOG_TABLE_NAME, "id", false);
         TableStatistics stats = TableStatistics.builder().setRowCount(Estimate.of(11.0)).build();
-        persistentStatistics.setTableStatistics(url, stats);
-        Optional<TableStatistics> fetchedStats = persistentStatistics.getTableStatistics(url);
+        persistentStatistics.setTableStatistics(table, stats);
+        Optional<TableStatistics> fetchedStats = persistentStatistics.getTableStatistics(table);
         assertEquals(stats, fetchedStats.orElseThrow());
     }
 
@@ -87,7 +113,6 @@ public class TestTrinoPersistentStatistics
     public void testTableStatisticsSerialization()
             throws JsonProcessingException
     {
-        initMocks(this);
         VastClient client = mockClient;
         VastConfig config = getMockServerReadyVastConfig();
         TrinoPersistentStatistics persistentStatistics = new TrinoPersistentStatistics(client, config);
@@ -97,14 +122,45 @@ public class TestTrinoPersistentStatistics
                 .setRowCount(Estimate.of(11.0))
                 .setColumnStatistics(colHandle, DUMMY_COLUMN_STATISTICS)
                 .build();
-        String url = "/buc/t";
+        VastTableHandle table = new VastTableHandle("buck/schem", "tab", "id", false);
         ObjectMapper mapper = TrinoStatisticsMapper.instance();
-        String tsBuffer = null;
-        tsBuffer = mapper.writeValueAsString(new TrinoSerializableTableStatistics(stats));
+        String tsBuffer = mapper.writeValueAsString(new TrinoSerializableTableStatistics(stats));
         when(mockClient.s3GetObj(anyString(), anyString()))
                 .thenReturn(Optional.ofNullable(tsBuffer));
-        persistentStatistics.deleteTableStatistics(url);
-        Optional<TableStatistics> newStats = persistentStatistics.getTableStatistics(url);
+        persistentStatistics.deleteTableStatistics(table);
+        Optional<TableStatistics> newStats = persistentStatistics.getTableStatistics(table);
         assertEquals(newStats.orElseGet(TableStatistics::empty), stats);
+    }
+
+    @Test
+    public void testTableStatisticsFetchRetryOnNetworkError()
+    {
+        VastConfig config = getMockServerReadyVastConfig();
+        VastClient spyClient = spy(new VastClient(null, config, new VastTrinoDependenciesFactory()));
+        TrinoPersistentStatistics persistentStatistics = new TrinoPersistentStatistics(spyClient, config, stats -> "{}");
+        VastTableHandle table = new VastTableHandle("buck/schem", "tab", "id", false);
+        try {
+            persistentStatistics.setTableStatistics(table, null);
+        }
+        catch (VastRuntimeException re) {
+            Throwable cause = re.getCause();
+            while (cause != null && cause.getCause() != null) {
+                cause = cause.getCause();
+            }
+            assertTrue(cause instanceof VastServerException, format("Cause is: %s", cause));
+        }
+        verify(spyClient, times(RETRY_MAX_COUNT + 1)).getPutObjectResult(anyString(), anyString(), anyString());
+
+        try {
+            persistentStatistics.getTableStatistics(table);
+        }
+        catch (VastRuntimeException re) {
+            Throwable cause = re.getCause();
+            while (cause != null && cause.getCause() != null) {
+                cause = cause.getCause();
+            }
+            assertTrue(cause instanceof VastServerException, format("Cause is: %s", cause));
+        }
+        verify(spyClient, times(RETRY_MAX_COUNT + 1)).getObjectAsString(anyString(), anyString());
     }
 }
